Revert "修复无法播放视频的问题"
