--- app/src/main/assets/playvideo.html
@@ -1,36 +1,0 @@
-<!DOCTYPE html>
-<!-- used by PlayYoukuVideoActivity -->
-<html lang="zh">
-
-<head>
-    <meta charset="utf-8"/>
-    <meta name="viewport" content="width=device-width, height=device-height, user-scalable=no"/>
-    <title>播放视频</title>
-    <style>
-        * {
-        padding: 0;
-        margin: 0;
-        }
-        html, body, #youkuplayer {
-        width: 100%;
-        height: 100%;
-        }
-    </style>
-</head>
-
-<body>
-<div id="youkuplayer"></div>
-<script type="text/javascript" src="http://player.youku.com/jsapi"></script>
-<script type="text/javascript">
-player = new YKU.Player('youkuplayer', {
-    styleid: webAppInterface.getStyleId(),
-    client_id: webAppInterface.getClientId(),
-    vid: webAppInterface.getVid(),
-    autoplay: webAppInterface.isAutoplay(),
-    show_related: webAppInterface.isShowRelated()
-});
-</script>
-
-</body>
-
-</html>
