+++ app/src/main/res/layout/activity_simple_web_view.xml
@@ -5,5 +5,5 @@
         android:id="@+id/web_view"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.PlayYoukuVideoActivity"
+        tools:context="com.optimalorange.cooltechnologies.ui.SimpleWebViewActivity"
         />
